The widened generator reaches the tokenizer intact
symbolic-data now hands over float64 (S6, 0a9302e). Two flash-ansr tests
compared a decoded span against float(np.float32(source)) -- a narrowing they
carried because the SERIALIZER used to narrow. It does not, so the span now
decodes to the source value exactly and the expectation was the only thing
still rounding.

Removing the np.float32() from those expectations makes them strictly
stronger: they assert the constant and the predict_y coordinate survive the
whole path -- generator, worker, collate, serializer, span -- bit for bit.

That is also the cross-repo integration check for S6: 542 of 544 flash-ansr
tests passed unchanged against the widened generator, and the two that did not
were both this stale expectation rather than a narrowing left in the product.

544 passed, 8 skipped. flake8 clean.

+++ tests/test_mask_modes.py
@@ -123,7 +123,9 @@ class TestPredictConstantsBlock:
                 # first span decodes to the first placeheld value, positional binding
                 first = block.index(IEEE754_START_TOKEN)
                 nibbles = block[first + 1:first + 1 + IEEE754_N_BYTES]
-                assert byte_tokens_to_float64(nibbles) == float(np.float32(draw["values"][0]))
+                # EXACT, not narrowed: the span carries the value at the precision it was
+                # fitted at. This read float(np.float32(...)) while the serializer narrowed.
+                assert byte_tokens_to_float64(nibbles) == draw["values"][0]
                 # loss discipline: openers force-fed, nibbles + closers supervised
                 mask = batch["task_mask"][row].tolist()
                 assert mask[start], "the block opener is harness-owned"
